+++ assets/sticker.R
@@ -1,0 +1,13 @@
+# img <- magick::image_read("assets/logo_base.png")
+#
+# hexSticker::sticker(
+#   img,
+#   package = "ukcpR",
+#   p_size = 15,
+#   p_color = "black",
+#   s_x = 1,
+#   s_y = 0.75,
+#   s_width = 0.6,
+#   h_fill = "#f0f0f0",
+#   filename = "logo.png"
+# )