+++ NEWS.md
@@ -1,3 +1,6 @@
 # ukcpR 0.1.0
 
-* TBC
+* Functions to make API calls for certain observational and projected .csv files.
+* Download the results of said calls locally.
+* Read and do some basic pre-processing of the .csv files in R.
+* Coordinate conversion from northings/eastings to latitude/longitude.